Adding permissions for the custom entities

+++ src/Form/MakeShoppingList.php
@@ -10,6 +10,7 @@ use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Session\AccountProxyInterface;
 use Drupal\recipes\Entity\ShoppingList;
 use Drupal\recipes\Entity\ShoppingListItem;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 /**
  * Implements an form to generate a Shopping List.
@@ -55,28 +56,29 @@ class MakeShoppingList extends FormBase
       $form['recipes'] = [
         '#type' => 'container',
         '#tree' => TRUE,
+        '#prefix' => '<div id="recipes-make-shopping-list">',
+        '#suffix' => '</div>',
       ];
 
       foreach($recipe_list->get('recipes')->referencedEntities() as $delta => $recipe) {
         
         $form['recipes'][$recipe->id()] = [
-          '#type' => 'details',
+          '#type' => 'fieldset',
           '#title' => $recipe->getTitle(),
           '#open' => TRUE
         ];
 
         foreach($recipe->get('field_recipes_ingredients')->referencedEntities() as $delta1 => $ingredient) {
           $ingredient_term = $ingredient->get('field_recipes_ingredient')->referencedEntities();
-          $label = array_filter([
-            $ingredient->get('field_recipes_ingredient_amount')->value ?: NULL,
-            reset($ingredient_term)->getName() ?: NULL,
-            $ingredient->get('field_recipes_ingredient_extra')->value ? '(' . $ingredient->get('field_recipes_ingredient_extra')->value . ')' : NULL,
-          ]);
 
           $form['recipes'][$recipe->id()][$ingredient->id()] = [
             '#type' => 'checkbox',
             '#default_value' => TRUE,
-            '#title' => implode(" ", $label),
+            '#title' => 'checkbox',
+            '#form_id' => $this->getFormId(),
+            '#amount' => $ingredient->get('field_recipes_ingredient_amount')->value ?: NULL,
+            '#ingredient' => reset($ingredient_term)->getName() ?: NULL,
+            '#extra' => $ingredient->get('field_recipes_ingredient_extra')->value ?: NULL
           ];
         }
       }
@@ -102,9 +104,20 @@ class MakeShoppingList extends FormBase
     ]);
     $shopping_list = reset($entities) ?: NULL;
     if ($shopping_list) {
+      // Make sure the user has permission to delete this shopping list.
+      if (!$shopping_list->access('delete', $this->current_user)) {
+        $this->messenger()->addError('You do not have permission to delete this shopping list.');
+        return;
+      }
       $shopping_list->delete();
     }
 
+    // Check permissions to create shopping lists.
+    $list_access_control_handler = $this->entity_type_manager->getAccessControlHandler('recipes_shopping_list');
+    if (!$list_access_control_handler->createAccess(NULL, $this->current_user)) {
+      $this->messenger()->addError('You do not have permission to create a new shopping list.');
+        return;
+    }
     // Create new shopping list.
     $shopping_list = ShoppingList::create([
       'name' => 'Shopping list for ' . $this->current_user->id(),
@@ -112,6 +125,13 @@ class MakeShoppingList extends FormBase
     ]);
     $shopping_list->save();
 
+    // Check permissions to create ShoppingListItems.
+    $list_item_access_control_handler = $this->entity_type_manager->getAccessControlHandler('recipes_shopping_list_item');
+    if (!$list_item_access_control_handler->createAccess(NULL, $this->current_user)) {
+      $this->messenger()->addError('You do not have permission to create a new shopping list item.');
+        return;
+    }
+    // Create all the items in the list.
     foreach($form_state->getValue('recipes') as $recipe_id => $ingredient_ids) {
       foreach($ingredient_ids as $ingredient_id => $checked) {
         if ($checked == 1) {
